+++ importers/src/Service/VendorService/BogPortalen/BogPortalenVendorService.php
@@ -10,8 +10,8 @@ use App\Exception\UnknownVendorServiceException;
 use App\Exception\UnsupportedIdentifierTypeException;
 use App\Service\VendorService\FtpDownloadService;
 use App\Service\VendorService\ProgressBarTrait;
-use App\Service\VendorService\SupportsSingleIdentifierInterface;
-use App\Service\VendorService\VendorServiceInterface;
+use App\Service\VendorService\VendorServiceImporterInterface;
+use App\Service\VendorService\VendorServiceSingleIdentifierInterface;
 use App\Service\VendorService\VendorServiceTrait;
 use App\Utils\CoverVendor\UnverifiedVendorImageItem;
 use App\Utils\Message\VendorImportResultMessage;
@@ -22,7 +22,7 @@ use Symfony\Component\Filesystem\Exception\FileNotFoundException;
 /**
  * Class BogPortalenVendorService.
  */
-class BogPortalenVendorService implements VendorServiceInterface, SupportsSingleIdentifierInterface
+class BogPortalenVendorService implements VendorServiceImporterInterface, VendorServiceSingleIdentifierInterface
 {
     use ProgressBarTrait;
     use VendorServiceTrait;